Syncing files in common folder, including moving Dispatcher into common folder.
