+++ src/js/actions/FacebookActions.js
@@ -1,5 +1,5 @@
 import { isWebApp } from '../utils/cordovaUtils'; // eslint-disable-line import/no-cycle
-import Dispatcher from '../dispatcher/Dispatcher';
+import Dispatcher from '../common/dispatcher/Dispatcher';
 import FacebookConstants from '../constants/FacebookConstants';
 import { oAuthLog } from '../utils/logging';
 import signInModalGlobalState from '../components/Settings/signInModalGlobalState';
