+++ src/js/actions/TwitterActions.js
@@ -1,4 +1,4 @@
-import Dispatcher from '../dispatcher/Dispatcher';
+import Dispatcher from '../common/dispatcher/Dispatcher';
 // import VoterActions from './VoterActions';
 // import VoterSessionActions from './VoterSessionActions';
