+++ src/js/actions/VoterActions.js
@@ -1,4 +1,4 @@
-import Dispatcher from '../dispatcher/Dispatcher';
+import Dispatcher from '../common/dispatcher/Dispatcher';
 
 let { hostname } = window.location;
 hostname = hostname || '';
