+++ src/js/common/actions/DonateActions.js
@@ -1,4 +1,4 @@
-import Dispatcher from '../../dispatcher/Dispatcher';
+import Dispatcher from '../dispatcher/Dispatcher';
 
 /*
 July 2021 TODO: Same named file in the WebApp and Campaigns -- PLEASE KEEP THEM IDENTICAL -- make symmetrical changes and test on both sides
