+++ src/js/common/actions/ShareActions.js
@@ -1,4 +1,4 @@
-import Dispatcher from '../../dispatcher/Dispatcher';
+import Dispatcher from '../dispatcher/Dispatcher';
 
 export default {
   emailRecipientListQueuedToSave (superShareItemId, emailToAdd, emailToRemove, resetEmailRecipientList) {
