+++ src/js/stores/VoterStore.js
@@ -1,6 +1,6 @@
 import { ReduceStore } from 'flux/utils';
 import cookies from '../utils/cookies';
-import Dispatcher from '../dispatcher/Dispatcher';
+import Dispatcher from '../common/dispatcher/Dispatcher';
 import { stringContains } from '../utils/textFormat';
